requirements.txt
geojson==3.0.1
numpy==1.23.5
openslide_python==1.2.0
pandas==2.0.2
Pillow==9.5.0
rasterio==1.3.7
shapely==2.0.1